requirements.txt
absl-py==2.1.0
accelerate==0.33.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
async-timeout==4.0.3
attrs==23.2.0
av==9.2.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.15.1
cfgv==3.4.0
charset-normalizer==3.3.2
cmake==3.29.3
coloredlogs==15.0.1
colorlog==6.8.2
datasets==2.20.0
diffusers==0.29.2
dill==0.3.8
distlib==0.3.8
exceptiongroup==1.2.2
expecttest==0.2.1
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-auth-oauthlib==0.4.6
greenlet==3.0.3
grpcio==1.64.0
habana-media-loader==1.16.0.526
habana-pyhlml==1.16.0.526
huggingface-hub==0.24.5
humanfriendly==10.0
identify==2.5.36
idna==3.7
importlib_metadata==8.2.0
iniconfig==2.0.0
Jinja2==3.1.4
joblib==1.4.2
Mako==1.3.5
Markdown==3.6
MarkupSafe==2.1.5
mpi4py==3.1.4
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
ninja==1.11.1.1
nodeenv==1.9.0
numpy==1.23.5
oauthlib==3.2.2
optimum==1.21.3
optuna==3.6.1
packaging==24.0
pandas==2.0.1
parameterized==0.9.0
pathspec==0.12.1
peft==0.12.0
perfetto==0.7.0
pillow==10.4.0
Pillow-SIMD==7.0.0.post3
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.3.3
protobuf==3.20.3
psutil==6.0.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.10.4
pycparser==2.22
pydantic==1.10.13
pynvml==8.0.4
pytest==7.4.4
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0
regex==2023.5.5
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
sentencepiece==0.2.0
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.32
symengine==0.11.0
sympy==1.12.1
tdqm==0.0.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.5.0
timm==1.0.8
tokenizers==0.19.1
tomli==2.0.1
torchsde==0.2.6
tqdm==4.66.4
trampoline==0.1.2
transformers==4.43.4
typing_extensions==4.12.0
tzdata==2024.1
urllib3==1.26.18
virtualenv==20.26.2
Werkzeug==3.0.3
xxhash==3.4.1
yamllint==1.35.1
yarl==1.9.4
zipp==3.19.2
